@@ -18,10 +18,12 @@ namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.Api
 {
     public class IntegrationFixtureBase
     {
+#pragma warning disable CS8618 // Initialized in Setup
         private IConfiguration _configuration;
         private TestServer _server;
         private IServiceScope _testScope;
         protected HttpClient Client;
+#pragma warning restore CS8618 //
 
         [SetUp]
         protected void Setup()
@@ -47,26 +49,22 @@ namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.Api
 
         private void CreateDatabase()
         {
-            using (IServiceScope scope = CreateScope())
-            {
-                var dbContext = scope.Resolve<BlueprintContext>();
-                // On Azure we cannot drop db, we can only delete all tables
-                DropAllDbObjects(dbContext.Database);
-                // In case that we want to delete db call: dbContext.Database.EnsureDeleted()
-                dbContext.Database.Migrate();
-            }
+            using IServiceScope scope = CreateScope();
+            var dbContext = scope.Resolve<BlueprintContext>();
+            // On Azure we cannot drop db, we can only delete all tables
+            DropAllDbObjects(dbContext.Database);
+            // In case that we want to delete db call: dbContext.Database.EnsureDeleted()
+            dbContext.Database.Migrate();
         }
 
         private void AddCurrentUserToDb()
         {
-            using (IServiceScope scope = CreateScope())
-            {
-                var currentUserProvider = scope.Resolve<ICurrentUserProvider>();
-                var dbContext = scope.Resolve<DbContext>();
+            using IServiceScope scope = CreateScope();
+            var currentUserProvider = scope.Resolve<ICurrentUserProvider>();
+            var dbContext = scope.Resolve<DbContext>();
 
-                dbContext.Add(currentUserProvider.User);
-                dbContext.SaveChanges();
-            }
+            dbContext.Add(currentUserProvider.User);
+            dbContext.SaveChanges();
         }
 
         [TearDown]
@@ -92,7 +90,7 @@ namespace Enigmatry.Blueprint.Api.Tests.Infrastructure.Api
                 string dropAllSql = EmbeddedResource.ReadResourceContent("Enigmatry.Blueprint.Api.Tests.Infrastructure.Database.DropAllSql.sql");
                 foreach (var statement in dropAllSql.SplitStatements())
                     // WriteLine("Executing: " + statement);
-                    database.ExecuteSqlCommand(statement);
+                    database.ExecuteSqlRaw(statement);
             }
             catch (SqlException ex)
             {
