@@ -7,10 +7,10 @@ namespace Enigmatry.Blueprint.Api.Models.Identity
     public class UserModel
     {
         public Guid Id { get; set; }
-        public string UserName { get; set; }
-        public string Name { get; set; }
+        public string UserName { get; set; } = String.Empty;
+        public string Name { get; set; } = String.Empty;
         public DateTimeOffset CreatedOn { get; set; }
         public DateTimeOffset UpdatedOn { get; set; }
-        public GitHubApi.User GitHubUser { get; set; }
+        public GitHubApi.User? GitHubUser { get; set; }
     }
 }
