@@ -1,11 +1,13 @@
-﻿namespace Enigmatry.Blueprint.Api.Models
+﻿using System;
+
+namespace Enigmatry.Blueprint.Api.Models
 {
     public class LocalizedMessagesModel
     {
-        public string Message1 { get; set; }
+        public string Message1 { get; set; } = String.Empty;
 
-        public string Message2 { get; set; }
+        public string Message2 { get; set; } = String.Empty;
         
-        public string Message3 { get; set; }
+        public string Message3 { get; set; } = String.Empty;
     }
 }
